changes for string value of annualpolicymaximumlimit field

+++ src/test/java/model/Response/Cart/GetCartRes/CoPay.java
@@ -6,8 +6,8 @@ import java.util.List;
 public class CoPay {
 
     private Double percentage;
-    private List<Integer> premium = null;
-    private List<Integer> taxAndServicePremium = null;
+    private List<Double> premium = null;
+    private List<Double> taxAndServicePremium = null;
 
     public Double getPercentage() {
         return percentage;
@@ -17,19 +17,19 @@ public class CoPay {
         this.percentage = percentage;
     }
 
-    public List<Integer> getPremium() {
+    public List<Double> getPremium() {
         return premium;
     }
 
-    public void setPremium(List<Integer> premium) {
+    public void setPremium(List<Double> premium) {
         this.premium = premium;
     }
 
-    public List<Integer> getTaxAndServicePremium() {
+    public List<Double> getTaxAndServicePremium() {
         return taxAndServicePremium;
     }
 
-    public void setTaxAndServicePremium(List<Integer> taxAndServicePremium) {
+    public void setTaxAndServicePremium(List<Double> taxAndServicePremium) {
         this.taxAndServicePremium = taxAndServicePremium;
     }
